requirements.txt
python-telegram-bot==20.5
yt-dlp==2025.8.1
